Correctly save 2D array with new interleave in envi.save_image [fixes #20].
When calling envi.save_image, a 2D numpy array is assumed to be a single-
band image. By default, the array is saved as a BIP image (for a single-
band image, the order of data in the image file is actually the same,
regardless of interleave). When an interleave other than "bip" was
specified, the function would fail because it assumed that the image data
would have three dimensions and the needed transpose could not be performed.