Do not set matplotlib interactive mode until a window is created (fixes #44).
The _init function was turning on mpl interactive mode as soon as spectral was
imported. This can change the behavior of mpl from client code that does not
use SPy's graphics capabilities. This was more likely to happen in code
running from the standard python interpreter (i.e., not ipython). Interactive
mode is now deferred until a window is graphical window is opened.

* spectral.py:
  (_init): Do not set matplotlib to interactive mode.
* graphics/spypylab.py:
  (ImageView.show): Call set_mpl_interactive.
  (imshow): Call set_mpl_interactive.
  (plot): Call set_mpl_interactive and import pyplot instead of pylab.
  (set_mpl_interactive): New function to ensure matplotlib is in interactive
   mode.