Set default BandInfo.band_{quantity,unit} to None (fixes #46).
When reading ENVI files, the band_quantity attribute of the SpyFile's BandInfo
object was being set to "Wavelength". While that is usually the case, it is
not always. Now, that assumption is no longer made and the  band_quantity and
band_unit attributes are defaulted to None in the BandInfo object (the previous
default was an empty string).

* graphics/spypylab.py:
  (plot): Handle band_quantity and band_unit default value None.
* io/envi.py:
  (open): band_quantity defaults to None instead of "Wavelength". If
    "wavelength units" is not specified in the header, default band_unit to
    None instead of empty string.
  (add_band_info_to_metadata): Handle and band_unit default value None.
* spectral.py:
  (BandInfo.__init__): Initialize band_quantity and band_unit to None.